#6 - design: SignupForm 컴포넌트 디자인 구조 수정
