+++ client/src/components/signup/SignupForm.jsx
@@ -3,33 +3,47 @@ import SignupBtn from "./SignupBtn";
 import GenderRadio from "./GenderRadio";
 import { useState } from "react";
 import ModalContainer from "./ModalCotainer";
-import SignupInput from "./SignupInput";
+import GuideModalContainer from "./GuideModalContainer";
+import {
+  Page,
+  CheckboxContent,
+  IdBlock,
+  CheckId,
+  CreatePwd,
+  CheckPwd,
+  CheckName,
+  CheckEmail,
+  Select,
+  CheckPhoneNum,
+  DetailAddress,
+  CheckAddress,
+  CheckBirth,
+} from "../../styles/signupStyle";
 
 function SignForm() {
   const [address, setAddress] = useState("");
   const [postAddress, setPostAddress] = useState("");
-  console.log(address);
 
   return (
     <Page>
       <IdBlock>
         <span>아이디</span>
         <span className="essential">*</span>
-        <SignupInput type={"id"} />
+        <CheckId></CheckId>
         <span>비밀번호</span>
-        <span className="essential">*</span>
-        <SignupInput type={"createPwd"} />
-        <SignupInput type={"checkPwd"} />
+        <span className="essential">*</span>{" "}
+        <CreatePwd placeholder="비밀먼호를 입력해주세요"></CreatePwd>
+        <CheckPwd placeholder="비밀번호를 확인"></CheckPwd>
         <span>이름</span> <span className="essential">*</span>
-        <SignupInput type={"name"} />
+        <CheckName placeholder="이름을 입력해주세요"></CheckName>
         <span>이메일</span>
         <span className="essential">*</span>
         <br></br>
-        <SignupInput type={"email"} />
+        <CheckEmail placeholder="이메일을 입력해주세요"></CheckEmail>
         <span>@</span>
         <Select>
           <option value="" hidden>
-            이메일을 선택해주세요
+            Type
           </option>
           <option value="1">gmail.com</option>
           <option value="2">naver.com</option>
@@ -38,7 +52,7 @@ function SignForm() {
         </Select>
         <span>휴대폰</span>
         <span className="essential">*</span>
-        <SignupInput type={"phoneNum"} />
+        <CheckPhoneNum placeholder="연락처를 입력해주세요"></CheckPhoneNum>
         <span>주소</span>
         <span className="essential">*</span>
         <br />
@@ -48,19 +62,14 @@ function SignForm() {
           onChange={(e) => setPostAddress(e.target.value)}
           placeholder="주소를 입력해주세요"
         ></CheckAddress>
-        {/* <SignupInput
-          type={"address"}
-          value={address}
-          onChange={(e) => setPostAddress(e.target.value)}
-        /> */}
         <ModalContainer setAddress={setAddress} />
-        <SignupInput type={"detailAddress"} />
-        <br />
+        <DetailAddress placeholder="나머지 주소를 입력해주세요"></DetailAddress>
         <span>생년월일</span>
+        <span className="essential">*</span>
         <br />
-        <SignupInput type={"birth"} />
-        <SignupInput type={"birth"} />
-        <SignupInput type={"birth"} />
+        <CheckBirth></CheckBirth>
+        <CheckBirth></CheckBirth>
+        <CheckBirth></CheckBirth>
         <br />
         <span>성별</span>
         <GenderRadio></GenderRadio>
@@ -83,81 +92,5 @@ function SignForm() {
     </Page>
   );
 }
-
-const Page = styled.div`
-  width: 100%
-  height: 100vh;
-  //   background-color: #f1f2f3;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  h2 {
-    font-size: 25px;
-    margin-bottom: 50px;
-  }
-`;
-
-const IdBlock = styled.div`
-  //   margin: 6px 0;
-  width: 100%;
-  .essential {
-    color: red;
-  }
-`;
-
-const CheckboxContent = styled.div`
-  display: flex;
-  align-items: flex-start;
-  width: 100%;
-  flex-direction: column;
-  .autoContent {
-    align-items: left;
-    margin-top: 8px;
-  }
-
-  .CheckboxText {
-    margin-left: 10px;
-    text-align: center;
-  }
-`;
-
-const Select = styled.select`
-  width: 40%;
-  min-height: 50px;
-  background: white;
-  //   color: gray;
-  padding-left: 5px;
-  font-size: 14px;
-
-  margin-left: 10px;
-
-  option {
-    color: black;
-    background: white;
-    display: flex;
-    white-space: pre;
-    min-height: 20px;
-    padding: 0px 2px 1px;
-  }
-`;
-
-const CheckAddress = styled.input.attrs({
-  type: "text",
-})`
-  display: inline-block;
-  width: 65%;
-  min-height: 50px;
-  padding: 7px 9px;
-  text-align: start;
-  font-size: 13px;
-  line-height: normal;
-  letter-spacing: normal;
-  box-shadow: none;
-  border-radius: 3px;
-  margin-top: 10px;
-  margin-bottom: 10px;
-`;
 
 export default SignForm;
